+++ app/src/androidTest/java/au/com/shiftyjelly/pocketcasts/models/db/AppDatabaseTest.kt
@@ -192,6 +192,7 @@ class AppDatabaseTest {
                 AppDatabase.MIGRATION_120_121,
                 AppDatabase.MIGRATION_121_122,
                 AppDatabase.MIGRATION_122_123,
+                AppDatabase.MIGRATION_123_124,
             )
             .build()
         // close the database and release any stream resources when the test finishes