+++ wear/src/main/kotlin/au/com/shiftyjelly/pocketcasts/wear/ui/WatchListScreen.kt
@@ -18,6 +18,7 @@ import au.com.shiftyjelly.pocketcasts.wear.ui.downloads.DownloadsScreen
 import au.com.shiftyjelly.pocketcasts.wear.ui.playlists.PlaylistsScreen
 import au.com.shiftyjelly.pocketcasts.wear.ui.podcasts.PodcastsScreen
 import au.com.shiftyjelly.pocketcasts.wear.ui.settings.SettingsScreen
+import au.com.shiftyjelly.pocketcasts.wear.ui.starred.StarredScreen
 import com.google.android.horologist.compose.layout.ScalingLazyColumn
 import com.google.android.horologist.compose.layout.ScalingLazyColumnState
 import au.com.shiftyjelly.pocketcasts.images.R as IR
@@ -106,6 +107,17 @@ fun WatchListScreen(
 
         item {
             WatchListChip(
+                title = stringResource(LR.string.profile_navigation_starred),
+                iconRes = IR.drawable.ic_starred,
+                onClick = {
+                    viewModel.onStarredClicked()
+                    navigateToRoute(StarredScreen.ROUTE)
+                },
+            )
+        }
+
+        item {
+            WatchListChip(
                 title = stringResource(LR.string.settings),
                 iconRes = IR.drawable.ic_profile_settings,
                 onClick = {
@@ -117,7 +129,7 @@ fun WatchListScreen(
     }
 }
 
-@Preview(device = WearDevices.SMALL_ROUND, showSystemUi = true)
+@Preview(device = WearDevices.SMALL_ROUND)
 @Composable
 private fun WatchListPreview() {
     WearAppTheme {